WIP logikk for småbarnstillegg

+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/autobrev/FinnAlleBarn6og18ÅrTask.kt
@@ -18,7 +18,7 @@ import java.time.LocalDate
     taskStepType = FinnAlleBarn6og18ÅrTask.TASK_STEP_TYPE,
     beskrivelse = "Send autobrev for barn som fyller 6 og 18 år til Dokdist",
     maxAntallFeil = 3,
-    triggerTidVedFeilISekunder = 60 * 60 * 24
+    triggerTidVedFeilISekunder = (60 * 60 * 24).toLong()
 )
 class FinnAlleBarn6og18ÅrTask(
     private val fagsakRepository: FagsakRepository,
@@ -40,7 +40,10 @@ class FinnAlleBarn6og18ÅrTask(
 
     private fun finnAlleBarnMedFødselsdagInneværendeMåned(alder: Long): Set<Fagsak> =
         LocalDate.now().minusYears(alder).let {
-            fagsakRepository.finnLøpendeFagsakMedBarnMedFødselsdatoInnenfor(it.førsteDagIInneværendeMåned(), it.sisteDagIMåned())
+            fagsakRepository.finnLøpendeFagsakMedBarnMedFødselsdatoInnenfor(
+                it.førsteDagIInneværendeMåned(),
+                it.sisteDagIMåned()
+            )
         }
 
     companion object {
